feat: 원화 달러 변환 추가

+++ spring-hello/src/main/java/com/springhello/domain/exchange/controller/ExchangeController.java
@@ -16,7 +16,7 @@ public class ExchangeController {
     private final ExchangeService exchangeService;
 
     @GetMapping("/exchange")
-    public Double exchange() {
+    public Float exchange() {
         return exchangeService.getExchangeRateUSD();
     }
 }
