+++ spring-hello/src/main/java/com/springhello/domain/product/entity/Product.java
@@ -8,16 +8,16 @@ public class Product {
 
     private final Long id;
     private final String name;
-    private final Double price;
+    private final Float price;
 
     @Builder
-    private Product(Long id, String name, Double price) {
+    private Product(Long id, String name, Float price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public static Product createProduct(Long id, String name, Double price) {
+    public static Product createProduct(Long id, String name, Float price) {
         return Product.builder()
                 .id(id)
                 .name(name)
